requirements.txt
langchain==0.2.12
langchain-core==0.2.29
langchain-openai==0.1.7
pinecone-client==3.2.2
openai==1.46.0
pymongo==4.6.3
python-dotenv==1.0.1
numpy==1.26.4
typing-extensions>=4.5.0
requests==2.32.4
pinecone==7.3.0